Move pytest fixtures etc. to a dedicated plugin
