+++ astrodata/tests/test_testing.py
@@ -3,76 +3,11 @@ Tests for the `astrodata.testing` module.
 """
 
 import os
-import subprocess
 
 import astrodata
 import numpy as np
 import pytest
-from astrodata.testing import (assert_same_class, download_from_archive,
-                               get_active_git_branch)
-
-
-def test_get_active_branch_name(capsys, monkeypatch):
-    """
-    Just execute and prints out the active branch name.
-    """
-
-    # With tracking branch
-    monkeypatch.setattr(subprocess, 'check_output',
-                        lambda *a, **k: b'(HEAD -> master, origin/master)')
-    assert get_active_git_branch() == 'master'
-    captured = capsys.readouterr()
-    assert captured.out == '\nRetrieved active branch name:  master\n'
-
-    # Only remote branch
-    monkeypatch.setattr(subprocess, 'check_output',
-                        lambda *a, **k: b'(HEAD, origin/foo)')
-    assert get_active_git_branch() == 'foo'
-
-    monkeypatch.setattr(subprocess, 'check_output',
-                        lambda *a, **k: b'(HEAD, origin/sky_sub, sky_sub)')
-    assert get_active_git_branch() == 'sky_sub'
-
-    # With other remote name
-    monkeypatch.setattr(subprocess, 'check_output',
-                        lambda *a, **k: b'(HEAD, myremote/foo)')
-    assert get_active_git_branch() == 'foo'
-
-    monkeypatch.setattr(subprocess, 'check_output',
-                        lambda *a, **k: b'(HEAD -> master, upstream/master)')
-    assert get_active_git_branch() == 'master'
-
-    # Raise error
-    def mock_check_output(*args, **kwargs):
-        raise subprocess.CalledProcessError(0, 'git')
-    monkeypatch.setattr(subprocess, 'check_output', mock_check_output)
-    assert get_active_git_branch() is None
-
-
-def test_change_working_dir(change_working_dir):
-    """
-    Test the change_working_dir fixture.
-
-    Parameters
-    ----------
-    change_working_dir : fixture
-        Custom DRAGONS fixture.
-    """
-    assert "astrodata/test_testing/outputs" not in os.getcwd()
-
-    with change_working_dir():
-        assert "astrodata/test_testing/outputs" in os.getcwd()
-
-    assert "astrodata/test_testing/outputs" not in os.getcwd()
-
-    with change_working_dir("my_sub_dir"):
-        assert "astrodata/test_testing/outputs/my_sub_dir" in os.getcwd()
-
-    assert "astrodata/test_testing/outputs" not in os.getcwd()
-
-    dragons_basetemp = os.getenv("$DRAGONS_TEST_OUT")
-    if dragons_basetemp:
-        assert dragons_basetemp in os.getcwd()
+from astrodata.testing import assert_same_class, download_from_archive
 
 
 def test_download_from_archive_raises_ValueError_if_envvar_does_not_exists():
@@ -147,15 +82,3 @@ def test_assert_same_class():
 
     with pytest.raises(AssertionError):
         assert_same_class(ad, np.array([1]))
-
-
-@pytest.mark.skip("Test fixtures might require some extra-work")
-def test_path_to_inputs(path_to_inputs):
-    assert isinstance(path_to_inputs, str)
-    assert "astrodata/test_testing/inputs" in path_to_inputs
-
-
-@pytest.mark.skip("Test fixtures might require some extra-work")
-def test_path_to_refs(path_to_refs):
-    assert isinstance(path_to_refs, str)
-    assert "astrodata/test_testing/refs" in path_to_refs
