+++ tox.ini
@@ -38,9 +38,7 @@ extras =
 deps =
     .jenkins/local_calibration_manager/GeminiCalMgr-0.9.11-py3-none-any.whl
 changedir =
-    {envsitepackagesdir}
-setenv =
-    PY_IGNORE_IMPORTMISMATCH = 1
+    .tmp/{envname}
 commands =
     python --version
     which python
